% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dea_brands.R
\docType{data}
\name{dea_brands}
\alias{dea_brands}
\title{Drug brand info}
\format{
A tibble with 28 rows and 2 variables:
\describe{
\item{category}{the drug category}
\item{brands}{Drug brand name}
}
}
\usage{
data(dea_brands)
}
\description{
A dataset containing a unique record for each the drug categories and their respective brand names.
}
\keyword{datasets}
