% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dea_controlled.R
\docType{data}
\name{dea_controlled}
\alias{dea_controlled}
\title{Drug controlled substance synonyms}
\format{
A tibble with 29 rows and 3 variables:
\describe{
\item{substance}{formal drug name}
\item{number}{DEA number}
\item{schedule}{drug schedule number}
\item{narcotic}{Yes No indicator if the drug is a narcotic}
\item{synonym}{synonym name}
}
}
\usage{
data(dea_controlled)
}
\description{
A dataset containing
}
\keyword{datasets}
