% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dea_street_names.R
\docType{data}
\name{dea_street_names}
\alias{dea_street_names}
\title{Drug slang from DEA's DIR-020-17 Drug Slang Code Words.pdf}
\format{
A tibble with 1734 rows and 3 variables:
\describe{
\item{class}{DEA drug class}
\item{brand}{Brand name of drug}
\item{slang}{slang name}
}
}
\usage{
data(dea_street_names)
}
\description{
A dataset containing slag provided by the DEA
}
\keyword{datasets}
