% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/drug_df.R
\docType{data}
\name{drug_df}
\alias{drug_df}
\title{Simulated drug data}
\format{
A tibble with 500 rows and 3 variables:
\describe{
\item{textdrug}{A free-text field of drug names}
\item{sex}{Patient sex}
\item{race}{Patient race}
}
}
\usage{
data(drug_df)
}
\description{
A dataset containing a record of drugs used by a patient. The \code{textdrug} field is unstructured.
}
\keyword{datasets}
