% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/drug_stop_words.R
\docType{data}
\name{drug_stop_words}
\alias{drug_stop_words}
\title{A vector with drug-specific stop words in English}
\format{
A vector with 47 words:
}
\usage{
data(drug_stop_words)
}
\description{
These are extraneous words or characters found in several clinical data that
have been accumulated to form a vector that can be used to parse out drug
names from corpus of text.
}
\keyword{datasets}
