% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/iqvia.R
\docType{data}
\name{iqvia}
\alias{iqvia}
\title{Data sent from IQVIA}
\format{
A tibble with 125 rows and 3 variables:
\describe{
\item{caegory}{drug category}
\item{class}{the drug class}
\item{synonym}{drug name}
}
}
\usage{
data(iqvia)
}
\description{
A dataset containing a unique record for each the drug names and their classes and categories.
}
\keyword{datasets}
