% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/lookup_df.R
\docType{data}
\name{lookup_df}
\alias{lookup_df}
\title{A lookup table with drug category, class, and synonyms}
\format{
A tibble with 4,296 rows and 3 variables:
\describe{
\item{category}{the drug category}
\item{class}{the drug class}
\item{synonym}{drug synonym}
}
}
\usage{
data(lookup_df)
}
\description{
A dataset containing a record for each drug synonym
}
\keyword{datasets}
