% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/lookup_syn.R
\name{lookup_syn}
\alias{lookup_syn}
\title{Make a table with the class and category for a drug name}
\usage{
lookup_syn(drug_name)
}
\arguments{
\item{drug_name}{a string of a single drug name.}
}
\value{
A lookup table with category \code{data.frame} having three columns:
drug class, drug category match, and synonym name.
}
\description{
This function provides a table with drug synonyms that have the
same class and category as the search term.
}
\examples{
  lookup_syn("zip")
}
