% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/noslang_raw.R
\docType{data}
\name{noslang_raw}
\alias{noslang_raw}
\title{Drug terms from https://www.noslang.com/drugs/dictionary}
\format{
A tibble with 3172 rows and 2 variables:
\describe{
\item{street_name}{Slang term}
\item{description}{Text description of drug term}
}
}
\source{
\href{https://www.noslang.com/drugs/dictionary}{www.noslang.com}
}
\usage{
data(noslang_raw)
}
\description{
A dataset containing a record for a drug rleated term and it's description.  This contains both drug names and other things like drug amounts.
}
\keyword{datasets}
