% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/noslang_street_names.R
\docType{data}
\name{noslang_street_names}
\alias{noslang_street_names}
\title{Drug slang from https://www.noslang.com/drugs/dictionary}
\format{
A tibble with 3172 rows and 2 variables:
\describe{
\item{street_name}{Slang term}
\item{description}{Text description of drug term}
}
}
\source{
\href{https://www.noslang.com/drugs/dictionary}{www.noslang.com}
}
\usage{
data(noslang_street_names)
}
\description{
A dataset containing a record for a drug slang term and it's description.
}
\keyword{datasets}
