\name{DOPE-package}
\alias{DOPE-package}
\alias{DOPE-package}
\docType{package}
\title{
\packageTitle{DOPE}
}
\description{
\packageDescription{DOPE}
}
\details{

The DESCRIPTION file:
\packageDESCRIPTION{DOPE}
\packageIndices{DOPE}

}
\author{
\packageAuthor{DOPE}

Maintainer: \packageMaintainer{DOPE}
}
\references{
 See \href{https://github.com/christophercschwarz/DOPE}{the GitHub page} for references. 
}

\keyword{ package }
\seealso{
 Primary Functions:
 \code{\link[DOPE]{DOPE}}
 \code{\link[DOPE]{RandomCorm}}
 \code{\link[DOPE]{plot_DOPE}}
 \code{\link[DOPE]{sensitivity_plot}}
}
