\name{DOPE-package}
\alias{DOPE-package}
\alias{DOPE-package}
\docType{package}
\title{
\packageTitle{Distributions of Possible Effects}
}
\description{
\packageDescription{Computational and graphical tools for sensitivity analysis with distibutions of possible effects}
}
\details{

The DESCRIPTION file:
\packageDESCRIPTION{DOPE}
\packageIndices{DOPE}

}
\author{
\packageAuthor{Christopher Schwarz}

Maintainer: \packageMaintainer{Christopher Schwarz}
}
\references{
 See \href{https://github.com/christophercschwarz/DOPE}{the GitHub page} for references. 
}

\keyword{ package }
\seealso{
 Primary Functions:
 \code{\link[DOPE]{DOPE}}
 \code{\link[DOPE]{RandomCorm}}
 \code{\link[DOPE]{plot_DOPE}}
 \code{\link[DOPE]{sensitivity_plot}}
}
