\name{DOPE-package}
\alias{DOPE-package}
\alias{DOPE}
\docType{package}
\title{
\packageTitle{DOPE}
}
\description{
\packageDescription{DOPE}
}
\details{

The DESCRIPTION file:
\packageDESCRIPTION{DOPE}
\packageIndices{DOPE}
 An overview of how to use the package, including the most important 
 functions 
}
\author{
\packageAuthor{DOPE}

Maintainer: \packageMaintainer{DOPE}
}
\references{
 Literature or other references for background information 
}

\keyword{ package }
\seealso{
 Optional links to other man pages, e.g. 
 \code{\link[<pkg>:<pkg>-package]{<pkg>}} 
}
\examples{
ex <- RandomCorm(5)
ex1 <- augment(ex)
}
