% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/parse.R
\name{parse}
\alias{parse}
\title{Parse a vector of free text containing drug information}
\usage{
parse(drug_vec)
}
\arguments{
\item{drug_vec}{A vector containing the free text to be parsed}
}
\value{
A n x 1 vector of class \code{character}.
}
\description{
This function provides a dataframe of parsed out strings from a
free text field, input as a vector, specified by the user.
}
\examples{
  parse("Lortab and Percocet")

}
