% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/parse.R
\name{parse}
\alias{parse}
\title{Parse a vector of free text containing drug information}
\usage{
parse(drug_vec)
}
\arguments{
\item{drug_vec}{A vector containing the free text to be parsed}
}
\value{
A n x 1 dataframe with class \code{data.frame}, \code{tbl_df}, \code{tbl}.
The dataframe has 1 variable, "drug"
}
\description{
This function provides a dataframe of parsed out strings from a
free text field given a file path specified by the user.
}
\examples{
\dontrun{
parse(drug_df)
}
}
